♻️ Refactor : 폴더 위치 적절한 위치로 변경
일부 폴더 위치 변경 및 불필요한 코드제거
