--- src/constants/url.ts
@@ -1,2 +1,0 @@
-export const YOUTUBE_API_BASED_URL = 'https://www.googleapis.com/youtube/v3';
-export const KAKAO_ROAD_BASED_URL = `https://dapi.kakao.com/v2/local/search/address.json?analyze_type=similar&query=`;
