+++ src/lib/types.ts
@@ -16,3 +16,15 @@ export interface relatedListItem {
   youtuber: string[];
   tag: string[];
 }
+
+export interface InputForm {
+  text?: string;
+  email: string;
+  password: string;
+  newpassword?: string;
+  passwordcheck: string;
+  nickname: string;
+  checkbox?: boolean;
+  file?: string;
+  verify?: string;
+}
