--- src/types/InputType.ts
@@ -1,11 +1,0 @@
-export interface InputForm {
-  text?: string;
-  email: string;
-  password: string;
-  newpassword?: string;
-  passwordcheck: string;
-  nickname: string;
-  checkbox?: boolean;
-  file?: string;
-  verify?: string;
-}
